+++ src/features/post/index.ts
@@ -1,1 +1,2 @@
-export { type PostWithAuthor } from "./model"
+export { type PostWithAuthor, useAddPostDialogStore, usePostsStore } from "./model"
+export { PostAddDialog } from "./ui"
