+++ src/features/post/model/index.ts
@@ -1,1 +1,3 @@
-export { type PostWithAuthor } from "./post-with-user"
+export { type PostWithAuthor } from "./post-with-author"
+export { useAddPostDialogStore } from "./use-add-post-dialog"
+export { usePostsStore } from "./use-posts-store"
