+++ src/features/post/model/use-add-post-dialog.ts
@@ -1,0 +1,16 @@
+import { create } from "zustand"
+
+interface AddPostDialogState {
+  showAddDialog: boolean
+}
+
+interface AddPostDialogActions {
+  setShowAddDialog: (showAddDialog: boolean) => void
+}
+
+type AddPostDialogStore = AddPostDialogState & AddPostDialogActions
+
+export const useAddPostDialogStore = create<AddPostDialogStore>((set) => ({
+  showAddDialog: false,
+  setShowAddDialog: (showAddDialog) => set({ showAddDialog }),
+}))
