+++ src/features/post/model/use-posts-store.ts
@@ -1,0 +1,18 @@
+import { create } from "zustand"
+
+import { PostWithAuthor } from "./post-with-author"
+
+interface PostsState {
+  posts: PostWithAuthor[]
+}
+
+interface PostsActions {
+  setPosts: (posts: PostWithAuthor[]) => void
+}
+
+type PostsStore = PostsState & PostsActions
+
+export const usePostsStore = create<PostsStore>((set) => ({
+  posts: [],
+  setPosts: (posts) => set({ posts }),
+}))
