+++ src/features/post/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as PostAddDialog } from "./PostAddDialog"
